Additional Skyboxes, Movement & Contrast using UI
Additional skyboxes have been added to allow movement in a loop rather than just end to end.

UI mentioned below is on the left hand.
Movement can be done by pressing a button on UI, this will be using a controller in future.
Contrast can be adjusted using a scrollbar in UI