Poke Input Only On Point, Image Names Adjusted
The AnimateHandOnInput class has been adjusted so that if either the grip or trigger value changes, these are checked to see if trigger == 0 and grip == 1, which represents a point animation. If so, set the poke object to active, else deactivate it. This means the menu can only be interacted with when pointing, which helps to prevent unwanted adjustments.

Skybox names have been adjusted so the numbering is always two digits. This will help in future when data is collected to know which skybox is activated by a user.