Basic JSON Data
Movement data is logged every time the user moves in the research level.

Movement data includes time taken at the current position, time taken in total, the coordinates moving from and to, and the number of moves so far.

Each set of movement data is added to a list. Currently a panel can be displayed which intends to show this data but is not working correctly.

The next step is to attempt to read from and write to files on the quest 2.