Wrist Watch Timer Added
The player has a wrist watch which will be used to display how much time they have left in the research level. Implementation is not complete yet as it will only display the time when being looked at, but this does not function correctly yet.

Moreover, the only functionality that it currently has is counting down. Nothing happens when the timer ends, and the timer is present in the tutorial level, which has no time limit.