Hand menu adjusted in Research Level
Contrast can now be changed within the research level, and the current location co-ordinates are displayed.

These additions are not pretty, and were simply for testing/information while deciding on a route to make for participants to take. Any functionalities which remain in future versions will be updated to actually look nice.