Wrist Watch Gaze Interaction Fixed
The wrist watch now appears and disappears depending on whether the player is looking at it or not, and does so independently of the hand menu UI. The affordance asset values needed to be changed, which I had forgotten to do. Also, the interaction layer masks need to be different for the hand UI and the wrist UI, although this is likely due to the way I have set up the gaze interaction, which is arguably back to front. This may be redesigned in future.