S-shape layout available
Added an "isResearch" bool which, when true, loads the grid with skyboxes only within the S-shape for research. This can be improved upon massively in future by removing the need for separate Material arrays for research and not for research, and simply eliminating specific skyboxes, but this is a very quick fix for now.