Movement Data Written to File in JSON Format
Move data was previously stored in JSON format but not written/saved anywhere as such. It was written to a file as a string but not using JSON. The SaveData.CreateSaveData() and SaveData.PrintSaveData() functions have been adjusted to ensure the text file contains movement data in a JSON format which can be easily read.

A future adjustment could be including the participant number/ID in the JSON data instead of having "NEW PARTICIPANT ****..." every few lines, making it even easier to be interpreted by a program reading the JSON data.