Contrast Moved to Debug UI in Research UI
Contrast was recently changed to be adjustable during the research level. This has been changed further to only be adjusted in the Unity editor as it is now part of the Debug UI (first introduced 20.08.2024 in commit "Debug UI Only in Editor, Start now at Beehive").

The font used for all text now has a slight black outline. This is mainly so that the timer is visible in brighter lighting. A panel was considered for this but was always out of scale. The feature may have to be re-built for the panel to work as intended.